moved TriggerableStage interface into its own file
