+++ orc/core/stages/cc_sink/cc_sink_stage.h
@@ -14,7 +14,7 @@
 #include "stage_parameter.h"
 #include <node_type.h>
 #include "video_field_representation.h"
-#include "../ld_sink/ld_sink_stage.h"  // For TriggerableStage interface
+#include "../triggerable_stage.h"
 #include <string>
 #include <memory>
 #include <functional>
