+++ orc/core/stages/chroma_sink/chroma_sink_stage.h
@@ -22,7 +22,7 @@
 #include <orc_source_parameters.h>
 #include "previewable_stage.h"
 #include "preview_renderer.h"  // For PreviewImage definition
-#include "../ld_sink/ld_sink_stage.h"  // For TriggerableStage interface
+#include "../triggerable_stage.h"
 
 #include <atomic>
 #include <thread>
